refactor: move query
